+++ plugins/woocommerce-admin/packages/components/src/tag/index.js
@@ -4,7 +4,8 @@
 import { __, sprintf } from '@wordpress/i18n';
 import { Fragment } from '@wordpress/element';
 import classnames from 'classnames';
-import { Button, Dashicon, Popover } from '@wordpress/components';
+import { Button, Popover } from '@wordpress/components';
+import { Icon, cancelCircleFilled } from '@wordpress/icons';
 import { decodeEntities } from '@wordpress/html-entities';
 import PropTypes from 'prop-types';
 import { withState, withInstanceId } from '@wordpress/compose';
@@ -86,7 +87,11 @@ const Tag = ( {
 					) }
 					aria-describedby={ labelId }
 				>
-					<Dashicon icon="dismiss" size={ 20 } />
+					<Icon
+						icon={ cancelCircleFilled }
+						size={ 20 }
+						className="clear-icon"
+					/>
 				</Button>
 			) }
 		</span>